LPS-66938 Move site teams data handlers to the proper packages
