+++ modules/apps/web-experience/site/site-teams-web/src/main/java/com/liferay/site/teams/web/internal/exportimport/data/handler/SiteTeamsPortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.site.teams.web.internal.lar;
+package com.liferay.site.teams.web.internal.exportimport.data.handler;
 
 import com.liferay.exportimport.kernel.lar.BasePortletDataHandler;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
